updated with a place to do funwave processing next to swash (play/development)
